Use of variables and comments
As it could be seen that variables in a makefile could be very useful
sometimes.It comes handy in situations where we want to change the
compiler or any options that go with it.

To use them we have to just assign the value to the variable before we
start writing targets. This commit demonstrates the use variables and
also comments.